requirements.txt
# minimum requirements for function execution.
firebase-admin==6.2.0
google-cloud==0.34.0
google-cloud-storage==2.10.0
numpy==1.23.5
Pillow==9.4.0
torchvision==0.15.0
onnxruntime==1.15.1